added final API routes

+++ Server/src/controllers/auth.ts
@@ -75,6 +75,8 @@ export const sendReVerificationToken: RequestHandler = async (req, res) => {
   const user = await User.findById(userId);
   if (!user) return res.status(403).json({ error: "Invalid request!" });
 
+  if (user.verified)return res.status(422).json({ error: "Your account is already verified!" });
+
   await EmailVerification.findOneAndDelete({
     owner: userId,
   });
